requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=2_gnu
anyio=3.6.1=py310hff52083_0
argon2-cffi=21.3.0=pyhd8ed1ab_0
argon2-cffi-bindings=21.2.0=py310h5764c6d_2
asgiref=3.5.2=pyhd8ed1ab_0
bcrypt=3.2.2=py310h5764c6d_0
brotlipy=0.7.0=py310h5764c6d_1004
bzip2=1.0.8=h7f98852_4
ca-certificates=2022.5.18.1=ha878542_0
certifi=2022.5.18.1=py310hff52083_0
cffi=1.15.0=py310h0fdd8cc_0
charset-normalizer=2.0.12=pyhd8ed1ab_0
click=8.1.3=py310hff52083_0
cryptography=37.0.2=py310h597c629_0
docopt=0.6.2=py_1
ecdsa=0.17.0=pyhd8ed1ab_0
fastapi=0.78.0=pyhd8ed1ab_0
flit-core=3.7.1=pyhd8ed1ab_0
greenlet=1.1.2=py310hd8f1fbe_2
h11=0.13.0=pyhd8ed1ab_1
idna=3.3=pyhd8ed1ab_0
keyutils=1.6.1=h166bdaf_0
krb5=1.19.3=h3790be6_0
ld_impl_linux-64=2.36.1=hea4e1c9_2
libedit=3.1.20191231=he28a2e2_2
libffi=3.4.2=h7f98852_5
libgcc-ng=12.1.0=h8d9b700_16
libgomp=12.1.0=h8d9b700_16
libnsl=2.0.0=h7f98852_0
libpq=14.3=hd77ab85_0
libstdcxx-ng=12.1.0=ha89aaad_16
libuuid=2.32.1=h7f98852_1000
libzlib=1.2.12=h166bdaf_0
ncurses=6.3=h27087fc_1
openssl=1.1.1o=h166bdaf_0
passlib=1.7.4=pyh9f0ad1d_0
pip=22.1.1=pyhd8ed1ab_0
pipreqs=0.4.11=pyhd8ed1ab_0
psycopg2=2.9.3=py310h6acc77f_0
pyasn1=0.4.8=py_0
pycparser=2.21=pyhd8ed1ab_0
pydantic=1.9.1=py310h5764c6d_0
pyopenssl=22.0.0=pyhd8ed1ab_0
pysocks=1.7.1=py310hff52083_5
python=3.10.4=h9a8a25e_0_cpython
python-jose=3.3.0=pyh6c4a22f_1
python-multipart=0.0.5=py_0
python_abi=3.10=2_cp310
readline=8.1=h46c0cb4_0
requests=2.27.1=pyhd8ed1ab_0
rsa=4.8=pyhd8ed1ab_0
setuptools=62.3.2=py310hff52083_0
six=1.16.0=pyh6c4a22f_0
sniffio=1.2.0=py310hff52083_3
sqlalchemy=1.4.36=py310h5764c6d_0
sqlite=3.38.5=h4ff8645_0
starlette=0.19.1=pyhd8ed1ab_0
tk=8.6.12=h27826a3_0
typing-extensions=4.2.0=hd8ed1ab_1
typing_extensions=4.2.0=pyha770c72_1
tzdata=2022a=h191b570_0
urllib3=1.26.9=pyhd8ed1ab_0
uvicorn=0.17.6=py310hff52083_1
wheel=0.37.1=pyhd8ed1ab_0
xz=5.2.5=h516909a_1
yarg=0.1.9=py_1
zlib=1.2.12=h166bdaf_0
